Fase 2 Double a BigDecimal

+++ backend/src/main/java/com/kristianconk/api_papeleria/proveedor/ProveedorResponseDTO.java
@@ -1,5 +1,7 @@
 package com.kristianconk.api_papeleria.proveedor;
 
+import java.math.BigDecimal;
+
 public record ProveedorResponseDTO(
         Long id,
         String nombre,
@@ -7,6 +9,6 @@ public record ProveedorResponseDTO(
         String telefono,
         String email,
         String contacto,
-        Double porcentajeComision
+        BigDecimal porcentajeComision
 ) {
 }
