+++ backend/src/main/java/com/kristianconk/api_papeleria/proveedor/ReporteVentasProveedorDTO.java
@@ -2,6 +2,7 @@ package com.kristianconk.api_papeleria.proveedor;
 
 import com.kristianconk.api_papeleria.ventas.VentaPorProductoDTO;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -11,9 +12,9 @@ public record ReporteVentasProveedorDTO(
         LocalDate fechaInicio,
         LocalDate fechaFin,
         Integer totalVentas,
-        Double montoTotalVentas,
-        Double comisionTienda,
-        Double montoAPagar,
+        BigDecimal montoTotalVentas,
+        BigDecimal comisionTienda,
+        BigDecimal montoAPagar,
         List<VentaPorProductoDTO> ventasPorProducto
 ) {
 }
